@@ -2,7 +2,7 @@ package io.techministry;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import io.techministry.ui.DashboardActivity;
 
